example/08_datastore/example_datastore_02_filter/requirements.txt
Flask==1.1.1
google-cloud-datastore==1.7.3
